Begin BitVector class for wider operations
The BitVector class will contain implementations of shits, wider bitwise
operations, and basic arithmetic.

Begin caching for constraint transformation: transformation results are
cached in case they are duplicated. At some point, this caching should
be configurable, but currently it is always performed. Perhaps it is
unnecessary, in which case, remove it later.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>