Introduce new methods in Model
With the new support for Python operators on a Variable, add solution
negation. We now track clauses we're waiting to add, letting us call
solver.add_clauses(...) once instead of multiple times, at the expense
of not being able to generate a CNF. Also adds helper for asserting a
variable is true (`add_assert`).

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>