Move cmsh/ to python/, add native to Makefile
This switches to using a pure-Makefile based build system. The intent
here is that build/ will eventually contain a working copy of the cmsh
module and library and an in-tree build of cryptominisat5, compiled how
we'd like it (C++ only, with M4RI / Gaussian elmination support).

The user will then be able to run:

    make all check && cd build && pip3 install --user -e .

and have a cmsh module available in their Python path.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>