Add cleanup routine for running instances
When run in a multiprocessing.Pool with several jobs, often a job will
terminate in a process, leaving the model and associated memory to be
garbage collected when Python feels like it. Allow the writer to free
most of the memory associated with a solver when they're done with it,
leaving only a little to be garbage collected.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>