Build updated solutions when adding clauses
Instead of building updated solutions all at once when querying a
missing variable, build the updated solution when the clause is added.
This redistributes the work to the construction phase and lessens it
by removing iterations through the clause tree.

Also, add a lval method to return the lbool value of a variable instead
of asserting like val does. This change has been pushed through to the
Python API, making it less brittle.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>