Note correct comparison to None
In our effort to override __eq__ and __ne__, we break code which
incorrectly does the following:

    if var == None:
        ...

or

    if vec == None:
        ...

Allow these (incorrect) usages to fail with a specific error message
telling the programmer to use is instead of == for checking None,
and for other comparisons to use the new equals() and not_equals()
methods.

These new methods will not fail with a type error, in order to behave
like __eq__ and __ne__ do.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>