Introduce mypy type checking
Begin the introduction of mypy type checking by adding a pyi stub for
the _native module. This allows mypy to find (and have type annotations)
for the object we expose from the native bindings. We also typecheck our
test cases. For this to pass, we remove the unused benchmark
benchmark_sum_array.py.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>