Add support for with-statement contexts
This change introduces the possibility of using the model in a with-as
statement:

    with cmsh.Model() as model:
        ...

This has the advantage of defining a clear scope for the model, and when
that scope is expired, automatically calling cleanup(). Note that calling
cleanup still remains optional, but is helpful when dealing with large
models on memory-constrained systems (where waiting for Python's GC to
free the underlying model might result in triggering the OOM killer).

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>